scripts/requirements-ingest.txt
google-cloud-aiplatform
psycopg2-binary
pgvector
markdown
beautifulsoup4
python-frontmatter
